+++ src/app/interfaces/INatures.ts
