+++ src/app/modals/generate-card-modal/generate-card-modal.component.html
