+++ src/app/modals/generate-card-modal/generate-card-modal.component.ts
