+++ src/app/pipes/stat-name.pipe.ts
