+++ src/app/pipes/stat-value.pipe.ts
