--- src/app/team-builder/team-builder-grid/team-builder-grid.component.scss
@@ -1,257 +1,0 @@
-/* ============================================
-            Types and Colors Key
-============================================*/
-$bug: #C3D21F;
-$electric: #FEE844;
-$fire: #FA5442;
-$grass: #86D04C;
-$normal: #B8B7A9;
-$rock: #CEBD72;
-$dark: #8D6855;
-$fairy: #F9ADFF;
-$flying: #79A3FF;
-$ground: #F0CF5C;
-$poison: #AA5DA1;
-$steel: #C4C2D9;
-$dragon: #8975FF;
-$fighting: #A75545;
-$ghost: #7975D7;
-$ice: #96F1FF;
-$psychic: #FA64B5;
-$water: #56B0FF;
-.bug {background-color: $bug} 
-.electric  {background-color: $electric} 
-.fire {background-color: $fire} 
-.grass {background-color: $grass} 
-.normal {background-color: $normal} 
-.rock {background-color: $rock} 
-.dark {background-color: $dark} 
-.fairy {background-color: $fairy} 
-.flying {background-color: $flying} 
-.ground {background-color: $ground} 
-.poison {background-color: $poison} 
-.steel {background-color: $steel} 
-.dragon {background-color: $dragon} 
-.fighting {background-color: $fighting} 
-.ghost {background-color: $ghost} 
-.ice {background-color: $ice} 
-.psychic {background-color: $psychic} 
-.water {background-color: $water}
-/* ============================================
-            Types and Colors Key
-============================================*/
-
-:host {
-    display: block;
-    margin: 1rem;
-}
-
-.team-container {
-    padding: 1rem;
-}
-
-.team-member:first-child {
-    margin-top: 0;
-}
-
-.team-member {
-    margin-top: 1rem;
-    border: 1px solid #CCCCCC;
-    display: grid;
-    grid-template-areas: 
-        "options"
-        "head"
-        "stats"
-        "finetune";
-    
-    .member-header {
-        width: calc(100% - 2rem);
-        padding: .5rem;
-        margin: 1rem;
-        display: flex;
-        height: 95% !important;
-        flex-direction: column;
-        align-content: center;
-        grid-area: head;
-    
-        h2 {
-            text-align: center;
-            margin: 0;
-        }
-    
-        img {
-            height: 120px;
-            width: 120px;
-            margin: 0 auto;
-        }
-    
-        div {
-            display: flex;
-            align-content: center;
-            flex-direction: row;
-            margin: .5rem;
-    
-            .type-icon {
-                display: inline-block;
-                color: white;
-                text-transform: uppercase;
-                font-size: 13px;
-                font-weight: 800;
-                padding: 3px;
-                margin: auto;
-                border: 3px solid white;
-                border-radius: 5px;
-                box-shadow: 0px 0px 4px 3px #000000;
-                text-shadow: 2px 2px 5px rgba(0,0,0,0.71);
-                border-radius: 17px;
-                font-family: 'Josefin Sans', sans-serif;
-            }
-        }
-    }
-    
-    .stats {
-        display: flex;
-        flex-direction: column;
-        width: 100%;
-        grid-area: stats;
-    
-        .stat {
-            display: flex;
-            margin: auto 0;
-            height: 2.5rem;
-            text-align: center;
-            align-content: space-between;
-            border: #CCCCCC solid 2px;
-    
-            span:first-child {
-                width: 15%;
-                margin: auto 0;
-            }
-    
-            span:last-child {
-                margin: auto 0;
-                padding: 5px;
-                background-color: #006D9A;
-                height: 100%;
-                text-align: left;
-                color: white;
-                font-weight: bolder;
-            }
-        }
-    }
-    
-    .pokemon-options {
-        width: 100%;
-        background-color: #CCCCCC;
-        display: flex;
-        flex-direction: row;
-        align-content: space-between;
-        grid-area: options;
-    
-        div {
-            margin: auto;
-            display: flex;
-            flex-direction: row;
-        }
-    
-        div button:nth-child(3) {
-            margin-left: 5px;
-        }
-    
-        button {
-            min-width: 35px;
-            width: 35px;
-            margin: 5px auto;
-    
-            hc-icon {
-                color: #000000;
-            }
-        }
-    
-        .giga-btn {
-            hc-icon {
-                color:#a082dd;
-            }
-        }
-    }
-    
-    hc-accordion {
-        grid-area: finetune;
-
-        hc-form-field {
-            display: flex;
-            flex-direction: column;
-            padding-right: 1rem !important;
-
-            hc-select {
-                margin: .5rem;
-            }
-        }
-    }
-}
-
-@media (min-width: 769px) {
-    .hc-accordion-trigger:focus {
-        box-shadow: none !important;
-    }
-    
-    hc-form-field {
-        flex-direction: row !important;
-        height: 3.5rem;
-
-        hc-select {
-            width: 15rem;
-        }
-    }
-}
-
-@media (min-width: 1100px) {
-    .team-member {
-        grid-template-columns: 20% 76% 4%;
-        grid-template-areas: 
-        "head stats options"
-        "finetune finetune options";
-
-        .member-header {
-            height: 100%;
-            padding: .5rem;
-            margin: 1rem;
-    
-            img {
-                margin: auto;
-            }
-        }
-    
-        .stats {
-            padding-right: .5rem;
-            margin-top: 1rem;
-        
-            .stat {
-                span:first-child {
-                    width: 10%;
-                }
-            }
-        }
-    
-        .pokemon-options {
-            flex-direction: column;
-    
-            button {
-                margin: auto;
-            }
-    
-            div {
-                flex-direction: column;
-            }
-    
-            div button {
-                margin: 5px auto;
-            }
-        
-            div button:nth-child(3) {
-                margin: 0 auto 5px auto;
-            }
-        }
-    }
-
-}
