--- src/app/team-builder/team-builder-menu/team-builder-menu.component.html
@@ -1,9 +1,0 @@
-<h1>Pokémon Team Builder</h1>
-
-<div class="builder-menu">
-    <button hc-button class="add-btn" buttonStyle="primary" [hcTooltip]="addTooltip">Add</button>
-    <button hc-button class="save-btn" buttonStyle="secondary" [hcTooltip]="saveTooltip">Save</button>
-    <button hc-button class="save-btn" buttonStyle="secondary" (click)="openLoadModal()">Load</button>
-    <button hc-button class="export-btn" buttonStyle="secondary" [hcTooltip]="exportTooltip">Export</button>
-    <button hc-button class="del-btn" buttonStyle="destructive">Delete Team</button>
-</div>
