--- src/app/team-builder/team-builder-menu/team-builder-menu.component.scss
@@ -1,23 +1,0 @@
-h1 {
-    margin: 65px 10px 0 10px;
-}
-
-.builder-menu {
-    margin: 0 10px;
-    
-    
-    button {
-        margin: 10px 10px 0 0;
-    }
-}
-
-@media (min-width: 769px) {
-
-}
-
-@media (min-width: 1100px) {
-    .builder-menu {
-        display: flex;
-    }
-    
-}
