+++ src/app/team-builder/team-builder.component.html
@@ -1,2 +1,58 @@
-<pkutil-team-builder-menu></pkutil-team-builder-menu>
-<pkutil-team-builder-grid></pkutil-team-builder-grid>
+<h1>Pokémon Team Builder</h1>
+
+<div class="builder-menu">
+    <button hc-button class="add-btn" buttonStyle="primary" [hcTooltip]="addTooltip">Add</button>
+    <button hc-button class="save-btn" buttonStyle="secondary" [hcTooltip]="saveTooltip">Save</button>
+    <button hc-button class="save-btn" buttonStyle="secondary" (click)="openLoadModal()">Load</button>
+    <button hc-button class="export-btn" buttonStyle="secondary" [hcTooltip]="exportTooltip">Export</button>
+    <button hc-button class="del-btn" buttonStyle="destructive">Delete Team</button>
+</div>
+
+<hc-tile class="team-container">
+    <div *ngFor = 'let pokemon of team' class="team-member" id="{{pokemon.guid}}">
+        <hc-tile class="member-header">
+            <h2>{{pokemon.name}}</h2>
+            <img src="{{pokemon.img}}" alt="Image of {{pokemon.name}}">
+            <div>
+                <p *ngFor = 'let type of pokemon.types' class="type-icon {{type}}">{{type}}</p>
+            </div>
+        </hc-tile>
+        <div class="stats" >
+            <div class="stat" *ngFor = 'let stat of pokemon.calcStats | keyvalue'>
+                <span>{{stat.key | statname }}</span>
+                <span style="width:{{stat.value | statvalue}}%">{{stat.value}}</span>
+            </div>
+        </div>
+        <div class="pokemon-options" >
+            <button hc-button (click)="deleteMember(pokemon.guid)" class="remove-btn" buttonStyle="destructive">
+                <hc-icon hcMenuIcon fontSet="fa" fontIcon="fa-trash"></hc-icon>
+            </button>
+            <div *ngIf="pokemon.megaData.canMegaEvo">
+                <button hc-button class="mega-btn" buttonStyle="primary-alt" [hcTooltip]="megaForm" *ngFor="let megaForm of pokemon.megaData.megaForms" name="peas">
+                    <img src="assets/mega-evo.webp" height="30px" alt="Mega Evolution Icon">
+                </button>
+            </div>
+            <button hc-button class="dyna-btn" buttonStyle="secondary" buttonStyle="pink" [hcTooltip]="dynaTooltip">
+                <img src="assets/dynamax.webp" height="30px" alt="Dynamax Icon">
+            </button>
+            <button hc-button class="giga-btn" buttonStyle="secondary" buttonStyle="pink" [hcTooltip]="gigaTooltip" *ngIf="pokemon.canGigantamax">
+                <hc-icon hcMenuIcon fontSet="fa" fontIcon="fa-cloud"></hc-icon>
+            </button>
+        </div>
+        <hc-accordion [triggerAlign]="alignment" #accordion>
+            <hc-accordion-toolbar>Fine Tune</hc-accordion-toolbar>
+            <hc-form-field>
+                <hc-select placeholder="Ability">
+                    <option *ngFor="let ability of pokemon.abilities">{{ability | abilityname}}</option>
+                </hc-select>
+                <hc-select placeholder="Nature" [formControl]="selectControl" (change)="updateStats()">
+                    <!-- THERE'S GOT TO BE A BETTER WAYYYYY -->
+                    <option *ngFor="let nature of natures" value="{{pokemon.guid}}::{{nature.name}}::{{nature.stats.attack}}::{{nature.stats.defense}}::{{nature.stats.speed}}::{{nature.stats.spattack}}::{{nature.stats.spdefense}}">{{nature.name | titlecase}}</option>
+                </hc-select>
+                <hc-select placeholder="Form" *ngIf="pokemon.forms.length > 0">
+                    <option *ngFor="let form of pokemon.forms">{{form}}</option>
+                </hc-select>
+            </hc-form-field>
+        </hc-accordion>
+    </div>
+</hc-tile>
